% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MixedDataImpute.R
\docType{package}
\name{MixedDataImpute}
\alias{MixedDataImpute}
\alias{MixedDataImpute-package}
\title{Missing data imputation for continuous and categorical data using nonparametric Bayesian joint models}
\description{
Missing data imputation for continuous and categorical data, using nonparametric Bayesian joint models (specifically the hierarchcially coupled mixture model with local dependence described in Murray and Reiter (2015); see citation(NPBayesMixedDataImpute)).
}

