% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare a dataset for imputation}
\usage{
prepare_data(X, Y, init = TRUE)
}
\arguments{
\item{X}{A data frame of categorical variables (as factors)}

\item{Y}{A matrix or data frame of continuous variables}

\item{init}{If \code{TRUE}, initialize missing values (from a marginal bootstrap of
observed values)}
}
\value{
An object of class \code{hcmm_data}
}
\description{
Prepares a dataset for imputation by mapping factor levels to integers and scaling
Y. Primarily used by \code{hcmm_impute} internally
}

