% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{remap_imputations}
\alias{remap_imputations}
\title{Map raw imputations back to original scale/factor labels.}
\usage{
remap_imputations(Ximp, Yimp, hcmmdat)
}
\arguments{
\item{Ximp}{A raw imputed X matrix}

\item{Yimp}{A raw imputed Y matrix}

\item{hcmmdat}{An \code{hcmm_data} object, used to recode/rescale the raw imputations}
}
\value{
A list with elements \code{X} and \code{Y} (transformed imputations).
}
\description{
Map raw imputations back to original scale (for continuous data) or factor labels.
Most users can ignore this function, which is primarily used by \code{hcmm_impute}
internally.
}

